requirements.txt
Django==5.0.6
sentence-transformers==2.2.2
torch==2.7.1
faiss-cpu==1.11.0
numpy==1.26.4
PyPDF2==3.0.0
google-api-python-client==2.90.0
google-generativeai==0.3.0
python-dotenv==1.1.1

